add aggregation api service
