+++ workspaces/marketplace/package.json
@@ -17,8 +17,8 @@
     "build:api-reports:only": "backstage-repo-tools api-reports -o ae-wrong-input-file-type,ae-undocumented --validate-release-tags",
     "build-image": "yarn workspace backend build-image",
     "clean": "backstage-cli repo clean",
-    "test": "backstage-cli repo test",
-    "test:all": "backstage-cli repo test --coverage",
+    "test": "backstage-cli repo test --detectOpenHandles",
+    "test:all": "backstage-cli repo test --coverage --detectOpenHandles",
     "test:e2e": "playwright test",
     "fix": "backstage-cli repo fix",
     "lint": "backstage-cli repo lint --since origin/main",
