+++ workspaces/marketplace/plugins/marketplace-common/src/api/MarketplaceCatalogClient.test.ts
@@ -13,14 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Knex } from 'knex';
 import { CatalogClient } from '@backstage/catalog-client';
 import { MarketplaceCatalogClient } from './MarketplaceCatalogClient';
-import {
-  AggregationsRequest,
-  MarketplaceApi,
-  MarketplaceKinds,
-} from '../types';
+import { MarketplaceKinds } from '../types';
 
 const mockPlugins = [
   {
@@ -62,21 +57,6 @@ beforeEach(() => {
   jest.clearAllMocks();
 });
 
-const getMockQueryBuilder = () => ({
-  whereNotNull: jest.fn().mockReturnThis(),
-  leftJoin: jest.fn().mockReturnThis(),
-  count: jest.fn().mockReturnThis(),
-  select: jest.fn().mockReturnThis(),
-  andWhereRaw: jest.fn().mockReturnThis(),
-  where: jest.fn().mockReturnThis(),
-  groupBy: jest.fn().mockReturnThis(),
-  having: jest.fn().mockReturnThis(),
-  orderBy: jest.fn().mockReturnThis(),
-});
-
-const getMockKnexClient = () =>
-  jest.fn(() => getMockQueryBuilder()) as unknown as jest.MockedFunction<Knex>;
-
 const options = {
   catalogApi: mockCatalogClient,
   auth: {
@@ -85,7 +65,6 @@ const options = {
       .mockResolvedValue('mockedServiceCredentials'),
     getPluginRequestToken: jest.fn().mockResolvedValue('mockedToken'),
   } as any,
-  client: getMockKnexClient(),
 };
 
 describe('MarketplaceCatalogClient', () => {
@@ -214,185 +193,6 @@ describe('MarketplaceCatalogClient', () => {
         'non-existent-featured-plugins',
       );
       expect(plugins).toHaveLength(0);
-    });
-  });
-
-  describe('getAggregateData', () => {
-    let mockKnexClient: jest.MockedFunction<Knex>;
-    let mockQueryBuilder: jest.Mocked<any>;
-    let api: MarketplaceApi;
-    beforeEach(() => {
-      mockQueryBuilder = getMockQueryBuilder();
-
-      mockKnexClient = jest.fn(() => ({
-        ...mockQueryBuilder,
-      })) as unknown as jest.MockedFunction<Knex>;
-
-      const mockRaw = jest
-        .fn()
-        .mockImplementation((sql: string) => sql) as unknown as Knex.RawBuilder<
-        any,
-        any
-      >;
-
-      mockKnexClient.raw = mockRaw;
-      api = new MarketplaceCatalogClient({
-        catalogApi: mockCatalogClient,
-        client: mockKnexClient,
-      });
-    });
-
-    it('should throw error if the database client is not passed', async () => {
-      const marketplaceCatalogApi = new MarketplaceCatalogClient({
-        ...options,
-        client: undefined as any,
-      });
-      await expect(marketplaceCatalogApi.getAggregateData([])).rejects.toThrow(
-        'Database client is not configured. Please check the MarketplaceCatalogClient configuration.',
-      );
-    });
-
-    it('should generate query for count aggregation', async () => {
-      const aggregationsQuery: AggregationsRequest = [
-        {
-          field: 'category.id',
-          type: 'count',
-          orderFields: [{ field: 'value', order: 'desc' }],
-        },
-      ];
-
-      await api.getAggregateData(aggregationsQuery);
-
-      expect(mockQueryBuilder.whereNotNull).toHaveBeenCalledWith(
-        'fe.final_entity',
-      );
-
-      expect(mockQueryBuilder.leftJoin).toHaveBeenCalledWith(
-        'search as s1',
-        'fe.entity_id',
-        's1.entity_id',
-      );
-
-      expect(mockQueryBuilder.count).toHaveBeenCalledWith('* as count');
-
-      expect(mockQueryBuilder.groupBy).toHaveBeenCalledWith(
-        's1.key',
-        's1.value',
-      );
-
-      expect(mockQueryBuilder.orderBy).toHaveBeenCalledWith('s1.value', 'desc');
-    });
-
-    it('should handle aggregations with filters and having clauses', async () => {
-      const aggregationsRequest: AggregationsRequest = [
-        {
-          type: 'sum',
-          field: 'amount',
-          name: 'total_amount',
-          filter: { category: 'electronics' },
-          havingFilter: { field: 'total_amount', operator: '>', value: '1000' },
-          orderFields: [{ field: 'value', order: 'asc' }],
-        },
-      ];
-
-      await api.getAggregateData(aggregationsRequest);
-
-      expect(mockQueryBuilder.leftJoin).toHaveBeenCalledWith(
-        'search as s1',
-        'fe.entity_id',
-        's1.entity_id',
-      );
-      expect(mockQueryBuilder.where).toHaveBeenCalledWith('s1.key', 'amount');
-      expect(mockQueryBuilder.leftJoin).toHaveBeenCalledWith(
-        'search as f1',
-        'fe.entity_id',
-        'f1.entity_id',
-      );
-      expect(mockQueryBuilder.where).toHaveBeenCalledWith('f1.key', 'category');
-      expect(mockQueryBuilder.where).toHaveBeenCalledWith(
-        'f1.value',
-        'electronics',
-      );
-
-      expect(mockQueryBuilder.select).toHaveBeenCalledWith(expect.anything());
-      expect(mockQueryBuilder.having).toHaveBeenCalledWith(
-        'total_amount',
-        '>',
-        '1000',
-      );
-
-      expect(mockQueryBuilder.orderBy).toHaveBeenCalledWith('s1.value', 'asc');
-    });
-
-    it('should handle having clauses with count', async () => {
-      const customBaseQuery = mockKnexClient('custom_table');
-      const aggregationsRequest: AggregationsRequest = [
-        {
-          field: 'kind',
-          type: 'count',
-          havingFilter: { field: 'count', operator: '>', value: '10' },
-        },
-      ];
-
-      await api.getAggregateData(aggregationsRequest, customBaseQuery);
-      expect(mockKnexClient).toHaveBeenCalledWith('custom_table');
-      expect(mockQueryBuilder.having).toHaveBeenCalledWith(
-        'COUNT(*)',
-        '>',
-        '10',
-      );
-    });
-
-    it('should use a custom base query if provided', async () => {
-      const customBaseQuery = mockKnexClient('custom_table');
-      const aggregationsRequest: AggregationsRequest = [
-        { type: 'count', field: 'custom_field' },
-      ];
-
-      await api.getAggregateData(aggregationsRequest, customBaseQuery);
-
-      expect(mockKnexClient).toHaveBeenCalledWith('custom_table');
-      expect(mockQueryBuilder.count).toHaveBeenCalledWith('* as count');
-    });
-
-    it('should handle orderFields', async () => {
-      const customBaseQuery = mockKnexClient('custom_table');
-      const aggregationsRequest: AggregationsRequest = [
-        {
-          field: 'kind',
-          value: 'plugin',
-          type: 'count',
-          orderFields: [{ field: 'count', order: 'asc' }],
-        },
-        {
-          field: 'label',
-          type: 'min',
-          orderFields: [{ field: 'value', order: 'desc' }],
-        },
-
-        {
-          field: 'name',
-          type: 'min',
-          orderFields: [{ field: 'incorrect-value' as any, order: 'desc' }],
-        },
-      ];
-
-      await api.getAggregateData(aggregationsRequest, customBaseQuery);
-
-      expect(mockKnexClient).toHaveBeenCalledWith('custom_table');
-
-      expect(mockQueryBuilder.orderBy).toHaveBeenCalledTimes(3);
-
-      const [field, order] = mockQueryBuilder.orderBy.mock.calls[0];
-      expect([field, order]).toEqual(['count', 'asc']);
-
-      const [labelField, labelFieldOrder] =
-        mockQueryBuilder.orderBy.mock.calls[1];
-      expect([labelField, labelFieldOrder]).toEqual(['s2.value', 'desc']);
-
-      const [nameField, nameFieldOrder] =
-        mockQueryBuilder.orderBy.mock.calls[2];
-      expect([nameField, nameFieldOrder]).toEqual(['s3.value', 'desc']);
     });
   });
 });
