+++ workspaces/marketplace/plugins/marketplace/src/plugin.ts
@@ -27,7 +27,12 @@ import {
 import MUIMarketplaceIcon from '@mui/icons-material/ShoppingBasketOutlined';
 
 import { rootRouteRef } from './routes';
-import { marketplaceApiRef, MarketplaceBackendClient } from './api';
+import {
+  MarketplaceAggregationApiRef,
+  MarketplaceAggregationBackendClient,
+  marketplaceApiRef,
+  MarketplaceBackendClient,
+} from './api';
 
 /**
  * Marketplace Plugin
@@ -47,6 +52,18 @@ export const marketplacePlugin = createPlugin({
       },
       factory: ({ discoveryApi, fetchApi }) =>
         new MarketplaceBackendClient({
+          discoveryApi,
+          fetchApi,
+        }),
+    }),
+    createApiFactory({
+      api: MarketplaceAggregationApiRef,
+      deps: {
+        discoveryApi: discoveryApiRef,
+        fetchApi: fetchApiRef,
+      },
+      factory: ({ discoveryApi, fetchApi }) =>
+        new MarketplaceAggregationBackendClient({
           discoveryApi,
           fetchApi,
         }),
